QueryFixAction: move back to ds types
